% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{abcd_cr}
\alias{abcd_cr}
\title{Calculates the monthly water balance using ABCD cold region model . 
This model work for cold region.}
\usage{
abcd_cr(
  para_a,
  para_b,
  para_c,
  para_d,
  SWS_ini,
  S_ini,
  G_ini,
  p,
  PE,
  temp,
  alpha,
  beta,
  Gmax
)
}
\arguments{
\item{para_a}{describes the amount of runoff and recharge in case the soils are under-saturated. This is dimensionless. This parameter range between 0.8-1}

\item{para_b}{his explains the saturation level of the soils. Unit of this variable is mm. This parameter values range between 20-160}

\item{para_c}{defines the ratio of groundwater recharge to surface runoff. This is dimensionless.This parameter values range lies between 0.8-1.}

\item{para_d}{dominants the rate of groundwater discharge. This is dimensionless.This parameter range between 0.001-0.01}

\item{SWS_ini}{is the solid water storage in the snow cover at the beginning of month. Unit of this variable is mm.}

\item{S_ini}{is the surface runoff}

\item{G_ini}{is the initial ground water storage}

\item{p}{is the precipitation}

\item{PE}{is the potential evaporation}

\item{temp}{is monthly temperature}

\item{alpha}{controlled by negative temperature. This unit is degree Celsius}

\item{beta}{is influence by negative temperature}

\item{Gmax}{is the potential maximum ground water storage}
}
\value{
abcd_cr function the output list of  a time-series of water balance components
}
\description{
Calculates the monthly water balance using ABCD cold region model . 
This model work for cold region.
}
\examples{
para_a=.5
para_b=200
para_c=1
para_d=.5
SWS_ini=20
S_ini=10
G_ini=100
alpha=0.03
beta=.143
Gmax=5
 m<-c()
m$p= c(100:111)
m$PE=c(50:61)
m$temp=c(-5:6)
dt<-data.table::as.data.table(m)
dt[, abcd_cr(para_a, para_b, para_c, para_d, SWS_ini, S_ini, G_ini, p, PE, temp, alpha, beta, Gmax)
 , by=.(p, PE, temp)]
}
