% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{day_cp}
\alias{day_cp}
\title{estimate the day using monthly time series of date .}
\usage{
day_cp(date)
}
\arguments{
\item{date}{is a monthly time series Date}
}
\value{
day_cp function the output is the day of date
}
\description{
estimate the day using monthly time series of date .
}
\examples{
 m<-c()
m$Date<-seq(as.Date("2014/1/1"), by="month" ,length.out =12)
dt<-data.table::as.data.table(m)
dt[, day:= day_cp(Date)]
dt
}
