% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{julianday_cp}
\alias{julianday_cp}
\title{Calculates the julian day using monthly time series of date .}
\usage{
julianday_cp(date)
}
\arguments{
\item{date}{is a monthly time series Date}
}
\value{
julianday_cp function the output of julian day vector.
}
\description{
Calculates the julian day using monthly time series of date .
}
\examples{
m<-c()
m$Date<-seq(as.Date("2014/1/1"), by="month" ,length.out =12)
dt<-data.table::as.data.table(m)
dt[, JD:= julianday_cp(Date)]
dt
}
