% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{month_cp}
\alias{month_cp}
\title{estimate the month using date time series data .}
\usage{
month_cp(date)
}
\arguments{
\item{date}{is a monthly time series Date}
}
\value{
month_cp function the output of the month vector.
}
\description{
estimate the month using date time series data .
}
\examples{
m<-c()
m$Date<-seq(as.Date("2014/1/1"), by="month" ,length.out =12)
dt<-data.table::as.data.table(m)
dt[, month:= month_cp(Date)]
dt
}
