% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pet_hamon}
\alias{pet_hamon}
\title{Calculates the potential evaporation using Hamon method . This method uses the input vector of  monthly date time series, temperature and longitude values.}
\usage{
pet_hamon(date, tavg, lat)
}
\arguments{
\item{date}{is a monthly time series Date}

\item{tavg}{is a monthly mean time series of temperature}

\item{lat}{is a vector of longitude}
}
\value{
pet_hamon function estimates the monthly potential evaporation in mm/day
}
\description{
Calculates the potential evaporation using Hamon method . This method uses the input vector of  monthly date time series, temperature and longitude values.
}
\examples{
 m<-c()
m$date<-seq(as.Date("2014/1/1"), by="month" ,length.out =12)
m$tavg<-c(3.1, 3.5, 5.0, 6.7, 9.3, 12.1, 14.3, 14.1, 11.8, 8.9, 5.5, 3.8)
m$lat<-rep(57.1526, 12)
 dt<-data.table::as.data.table(m)
dt[, PET_Hamon:= pet_hamon(date, tavg, lat)]
dt
}
