% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{year_cp}
\alias{year_cp}
\title{Calculates the year using monthly time series of date .}
\usage{
year_cp(date)
}
\arguments{
\item{date}{is a monthly time series Date}
}
\value{
year_cp function the output of the year vector.
}
\description{
Calculates the year using monthly time series of date .
}
\examples{
m<-c()
m$Date<-seq(as.Date("2014/1/1"), by="month" ,length.out =12)
dt<-data.table::as.data.table(m)
dt[, year:= year_cp(Date)]
dt
}
